+++ aql-engine/src/test/java/org/ehrbase/openehr/aqlengine/asl/AslGraph.java
@@ -48,6 +48,7 @@ import org.ehrbase.openehr.aqlengine.asl.model.field.AslField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslOrderByField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslSubqueryField;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslAuditDetailsJoinCondition;
+import org.ehrbase.openehr.aqlengine.asl.model.join.AslCommitterJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslDelegatingJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoin;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoinCondition;
@@ -221,6 +222,10 @@ public class AslGraph {
                     case AslDelegatingJoinCondition c -> "DelegatingJoinCondition %s ->\n%s"
                             .formatted(c.getLeftOwner().getAlias(), conditionToGraph(level + 2, c.getDelegate()));
                     case AslAuditDetailsJoinCondition c -> "AuditDetailsJoinCondition %s -> %s"
+                            .formatted(
+                                    c.getLeftOwner().getAlias(),
+                                    c.getRightOwner().getAlias());
+                    case AslCommitterJoinCondition c -> "CommitterJoinCondition %s -> %s"
                             .formatted(
                                     c.getLeftOwner().getAlias(),
                                     c.getRightOwner().getAlias());