+++ rm-db-format/src/main/java/org/ehrbase/openehr/dbformat/VersionedObjectDataStructure.java
@@ -98,8 +98,8 @@ public final class VersionedObjectDataStructure {
     public static void fillInMagnitudes(JsonNode jsonNode) {
         streamObjectNodes(jsonNode)
                 .filter(JsonNode::isObject)
-                .forEach(n ->
-                        addMagnitudeAttribute(n.get(DbToRmFormat.TYPE_ATTRIBUTE).textValue(), n));
+                .forEach(n -> addMagnitudeAttribute(
+                        n.get(RmAttribute.OBJ_TYPE.attribute()).textValue(), n));
     }
 
     private static void addMagnitudeAttribute(String type, ObjectNode object) {
@@ -258,7 +258,7 @@ public final class VersionedObjectDataStructure {
     private static Optional<String> getType(JsonNode childNode) {
         return Optional.of(childNode)
                 .filter(JsonNode::isObject)
-                .map(n -> n.get(DbToRmFormat.TYPE_ATTRIBUTE))
+                .map(n -> n.get(RmAttribute.OBJ_TYPE.attribute()))
                 .map(JsonNode::asText);
     }
 
@@ -266,7 +266,7 @@ public final class VersionedObjectDataStructure {
         ObjectNode newNode = jsonNode.objectNode();
 
         jsonNode.fields().forEachRemaining(e -> {
-            String alias = RmAttributeAlias.getAlias(e.getKey());
+            String alias = RmAttribute.getAlias(e.getKey());
             JsonNode child = e.getValue();
             if (child.isObject()) {
                 child = applyRmAliases((ObjectNode) child);
@@ -280,8 +280,8 @@ public final class VersionedObjectDataStructure {
                     }
                 });
                 child = newArray;
-            } else if (child.isTextual() && e.getKey().equals(DbToRmFormat.TYPE_ATTRIBUTE)) {
-                String rmNameAlias = RmTypeAlias.getAlias(child.textValue());
+            } else if (child.isTextual() && e.getKey().equals(RmAttribute.OBJ_TYPE.attribute())) {
+                String rmNameAlias = RmType.getAlias(child.textValue());
                 child = new TextNode(rmNameAlias);
             }
             newNode.putIfAbsent(alias, child);